refactor: convert remaining frame-based logic to time-based
Convert all internal logic to work with time directly instead of
converting between frames and time.

Changes:

Player (src/ofxAEPlayer.cpp, src/ofxAEPlayer.h):
- target_frame_ → target_time_ (float → double seconds)
- constrainFrame() → constrainTime() (unified time-based constraint)
- updatePlayback(): Direct time arithmetic for all loop modes
- getPosition()/setPosition(): Direct time calculations
- nextFrame()/previousFrame(): Time-based stepping

Layer (src/core/ofxAELayer.cpp, src/core/ofxAELayer.h):
- in_/out_ → in_time_/out_time_ (int frames → double seconds)
- getInPoint()/getOutPoint(): Return time directly
- getDuration(): Direct time subtraction
- isActiveAtTime(): Direct time comparison

Benefits:
- Eliminates frame-to-time conversion overhead in hot paths
- More precise calculations (no frame quantization)
- Simpler code (fewer intermediate variables)
- No cumulative rounding errors

Remaining frame-based code:
- Public API setFrame()/getCurrentFrame() (compatibility wrappers)
- Image sequence indexing (frame number needed for file naming)
- JSON import (one-time conversion from AE's frame-based export)

All internal timing now operates on seconds. Frame numbers are
used only at API boundaries or where truly necessary (file indexing).

Documentation: FRAME_TO_TIME_CONVERSION_SUMMARY.md

+++ src/core/ofxAEComposition.h
@@ -22,7 +22,7 @@ public:
 		int height;
 		struct LayerInfo {
 			std::string name, unique_name, filepath, parent;
-			int offset;
+			float offset;
 			bool visible;
 			struct TrackMatte {
 				std::string layer;
@@ -62,7 +62,7 @@ private:
 	std::vector<std::shared_ptr<Layer>> layers_;
 	std::map<std::string, std::weak_ptr<Layer>> name_layers_map_;
 	std::map<std::string, std::weak_ptr<Layer>> unique_name_layers_map_;
-	std::map<std::weak_ptr<Layer>, int, std::owner_less<std::weak_ptr<Layer>>> layer_offsets_;
+	std::map<std::weak_ptr<Layer>, float, std::owner_less<std::weak_ptr<Layer>>> layer_offsets_;
 
 	double current_time_;
 };
